+++ j2k/src/org/jetbrains/kotlin/j2k/ReferenceSearcher.kt
@@ -17,6 +17,7 @@
 package org.jetbrains.kotlin.j2k
 
 import com.intellij.psi.*
+import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.PsiUtil
 
 interface ReferenceSearcher {
@@ -66,6 +67,13 @@ fun PsiField.isVar(searcher: ReferenceSearcher): Boolean {
 
 fun PsiVariable.hasWriteAccesses(searcher: ReferenceSearcher, scope: PsiElement?): Boolean
         = if (scope != null) searcher.findVariableUsages(this, scope).any { PsiUtil.isAccessedForWriting(it) } else false
+
+fun PsiVariable.isInVariableInitializer(searcher: ReferenceSearcher, scope: PsiElement?): Boolean {
+    return if (scope != null) searcher.findVariableUsages(this, scope).any {
+        val parent = PsiTreeUtil.skipParentsOfType(it, PsiParenthesizedExpression::class.java)
+        parent is PsiVariable && parent.initializer == it
+    } else false
+}
 
 object EmptyReferenceSearcher: ReferenceSearcher {
     override fun findLocalUsages(element: PsiElement, scope: PsiElement): Collection<PsiReference> = emptyList()