+++ j2k/testData/fileOrElement/field/specifyType.kt
@@ -26,4 +26,41 @@ internal class A {
         field9 = null
         field10 = null
     }
+
+    internal interface I
+
+    private val anonymous: I = object : I {
+
+    }
+
+    var anonymous2: I = object : I {
+
+    }
+
+    private var anonymous3: I = object : I {
+
+    }
+
+    private var iimpl = anonymous
+
+    fun testAnonymousObject(i: Any) {
+        if (true) {
+            iimpl = i as I
+        } else if (true) {
+            anonymous3 = i as I
+        }
+
+        val anonymousLocal1: I = object : I {
+
+        }
+
+        var anonymousLocal2: I = object : I {
+
+        }
+
+        val iimpl = anonymousLocal1
+        if (true) {
+            anonymousLocal2 = i as I
+        }
+    }
 }